+++ packages/xchain-cosmos/src/util.ts
@@ -1,14 +1,10 @@
 import { cosmosclient, proto } from '@cosmos-client/core'
-import { FeeType, Fees, Tx, TxFrom, TxTo, TxType } from '@xchainjs/xchain-client'
-import { Asset, assetToString, baseAmount } from '@xchainjs/xchain-util'
+import { FeeType, Fees, Network, Tx, TxFrom, TxTo, TxType } from '@xchainjs/xchain-client'
+import { Asset, BaseAmount, CosmosChain, baseAmount, eqAsset } from '@xchainjs/xchain-util'
 
-import { APIQueryParam, RawTxResponse, TxResponse } from './cosmos/types'
-import { AssetAtom, AssetMuon } from './types'
-
-/**
- * The decimal for cosmos chain.
- */
-export const DECIMAL = 6
+import { AssetAtom, DECIMAL } from './const'
+import { APIQueryParam, TxResponse } from './cosmos/types'
+import { ChainIds, ClientUrls as ClientUrls } from './types'
 
 /**
  * Type guard for MsgSend
@@ -38,8 +34,7 @@ export const isMsgMultiSend = (msg: unknown): msg is proto.cosmos.bank.v1beta1.M
  * @returns {string} The denomination of the given asset.
  */
 export const getDenom = (asset: Asset): string => {
-  if (assetToString(asset) === assetToString(AssetAtom)) return 'uatom'
-  if (assetToString(asset) === assetToString(AssetMuon)) return 'umuon'
+  if (eqAsset(asset, AssetAtom)) return 'uatom'
   return asset.symbol
 }
 
@@ -51,32 +46,64 @@ export const getDenom = (asset: Asset): string => {
  */
 export const getAsset = (denom: string): Asset | null => {
   if (denom === getDenom(AssetAtom)) return AssetAtom
-  if (denom === getDenom(AssetMuon)) return AssetMuon
+  // IBC assets
+  if (denom.startsWith('ibc/'))
+    // Note: Don't use `assetFromString` here, it will interpret `/` as synth
+    return {
+      chain: CosmosChain,
+      symbol: denom,
+      // TODO (xchain-contributors)
+      // Get readable ticker for IBC assets from denom #600 https://github.com/xchainjs/xchainjs-lib/issues/600
+      // At the meantime ticker will be empty
+      ticker: '',
+      synth: false,
+    }
   return null
 }
 
-const getCoinAmount = (coins?: proto.cosmos.base.v1beta1.ICoin[]) => {
-  return coins
-    ? coins
-        .map((coin) => baseAmount(coin.amount || 0, DECIMAL))
-        .reduce((acc, cur) => baseAmount(acc.amount().plus(cur.amount()), DECIMAL), baseAmount(0, DECIMAL))
-    : baseAmount(0, DECIMAL)
-}
 /**
- * Parse transaction type
+ * Parses amount from `ICoin[]`
+ *
+ * @param {ICoin[]} coinst List of coins
+ *
+ * @returns {BaseAmount} Coin amount
+ */
+const getCoinAmount = (coins: proto.cosmos.base.v1beta1.ICoin[]): BaseAmount =>
+  coins
+    .map((coin) => baseAmount(coin.amount || 0, DECIMAL))
+    .reduce((acc, cur) => baseAmount(acc.amount().plus(cur.amount()), DECIMAL), baseAmount(0, DECIMAL))
+
+/**
+ * Filters `ICoin[]` by given `Asset`
+ *
+ * @param {ICoin[]} coinst List of coins
+ * @param {Asset} asset Asset to filter coins
+ *
+ * @returns {ICoin[]} Filtered list
+ */
+const getCoinsByAsset = (coins: proto.cosmos.base.v1beta1.ICoin[], asset: Asset): proto.cosmos.base.v1beta1.ICoin[] =>
+  coins.filter(({ denom }) => {
+    const coinAsset = !!denom ? getAsset(denom) : null
+    return !!coinAsset ? eqAsset(coinAsset, asset) : false
+  })
+
+/**
+ * Parses transaction history
  *
  * @param {TxResponse[]} txs The transaction response from the node.
- * @param {Asset} mainAsset Current main asset which depends on the network.
- * @returns {Tx[]} The parsed transaction result.
+ * @param {Asset} asset Asset to get history of transactions from
+ *
+ * @returns {Tx[]} List of transactions
  */
-export const getTxsFromHistory = (txs: TxResponse[], mainAsset: Asset): Tx[] => {
+export const getTxsFromHistory = (txs: TxResponse[], asset: Asset): Tx[] => {
   return txs.reduce((acc, tx) => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let msgs: Array<proto.cosmos.bank.v1beta1.MsgSend | proto.cosmos.bank.v1beta1.MsgMultiSend>
-    if ((tx.tx as RawTxResponse).body === undefined) {
-      msgs = cosmosclient.codec.packAnyFromCosmosJSON(tx.tx).msg
-    } else {
-      msgs = cosmosclient.codec.packAnyFromCosmosJSON((tx.tx as RawTxResponse).body.messages)
+    let msgs: Array<proto.cosmos.bank.v1beta1.MsgSend | proto.cosmos.bank.v1beta1.MsgMultiSend> = []
+    if (tx.tx) {
+      if (!tx.tx.body) {
+        msgs = cosmosclient.codec.packAnyFromCosmosJSON(tx.tx).msg
+      } else {
+        msgs = cosmosclient.codec.packAnyFromCosmosJSON(tx.tx.body.messages)
+      }
     }
 
     const from: TxFrom[] = []
@@ -84,7 +111,8 @@ export const getTxsFromHistory = (txs: TxResponse[], mainAsset: Asset): Tx[] => 
     msgs.map((msg) => {
       if (isMsgSend(msg)) {
         const msgSend = msg
-        const amount = getCoinAmount(msgSend.amount)
+        const coins = getCoinsByAsset(msgSend.amount, asset)
+        const amount = getCoinAmount(coins)
 
         let from_index = -1
 
@@ -119,7 +147,8 @@ export const getTxsFromHistory = (txs: TxResponse[], mainAsset: Asset): Tx[] => 
         const msgMultiSend = msg
 
         msgMultiSend.inputs.map((input) => {
-          const amount = getCoinAmount(input.coins || [])
+          const coins = getCoinsByAsset(input.coins || [], asset)
+          const amount = getCoinAmount(coins)
 
           let from_index = -1
 
@@ -138,7 +167,8 @@ export const getTxsFromHistory = (txs: TxResponse[], mainAsset: Asset): Tx[] => 
         })
 
         msgMultiSend.outputs.map((output) => {
-          const amount = getCoinAmount(output.coins || [])
+          const coins = getCoinsByAsset(output.coins || [], asset)
+          const amount = getCoinAmount(coins)
 
           let to_index = -1
 
@@ -161,7 +191,7 @@ export const getTxsFromHistory = (txs: TxResponse[], mainAsset: Asset): Tx[] => 
     return [
       ...acc,
       {
-        asset: mainAsset,
+        asset,
         from,
         to,
         date: new Date(tx.timestamp),
@@ -206,3 +236,35 @@ export const getDefaultFees = (): Fees => {
  *
  **/
 export const getPrefix = () => 'cosmos'
+
+/**
+ * Default client urls
+ *
+ * @returns {ClientUrls} The client urls for Cosmos.
+ */
+export const getDefaultClientUrls = (): ClientUrls => {
+  const mainClientUrl = 'https://api.cosmos.network'
+  // Note: In case anyone facing into CORS issue, try the following URLs
+  // https://lcd-cosmos.cosmostation.io/
+  // https://lcd-cosmoshub.keplr.app/
+  // @see (Discord #xchainjs) https://discord.com/channels/838986635756044328/988096545926828082/988103739967688724
+  return {
+    [Network.Testnet]: 'https://rest.sentry-02.theta-testnet.polypore.xyz',
+    [Network.Stagenet]: mainClientUrl,
+    [Network.Mainnet]: mainClientUrl,
+  }
+}
+
+/**
+ * Default chain ids
+ *
+ * @returns {ChainIds} Chain ids for Cosmos.
+ */
+export const getDefaultChainIds = (): ChainIds => {
+  const mainChainId = 'cosmoshub-4'
+  return {
+    [Network.Testnet]: 'theta-testnet-001',
+    [Network.Stagenet]: mainChainId,
+    [Network.Mainnet]: mainChainId,
+  }
+}