+++ packages/xchain-cosmos/CHANGELOG.md
@@ -1,3 +1,32 @@
+# v.0.18.0-alpha.1 (2022-06-20)
+
+## Add
+
+- const `DEFAULT_GAS_LIMIT`
+- const `DEFAULT_FEE`
+- Optional parameter `gasLimit` in `transfer` and `transferOffline`
+- Params `clientUrls`, `chainIds` for constructor
+- Helper `getDefaultClientUrls`
+- Helper `getDefaultChainIds`
+- Setter `setNetwork`
+
+## Fix
+
+- `getFees` returns values based on `DEFAULT_FEE`
+- Initial one instance of `CosmosSDKClient` only depending on network
+- Support IBC assets in `getBalances` (#596)
+- Get IBC assets from denom in `getAsset`
+
+## Update
+
+- Result of `getTxsFromHistory` is filtered by given asset
+- Move misc. constants into `const.ts`
+
+## Breaking change
+
+- Remove deprecated `AssetMuon`
+- Remove deprecated `Client.getMainAsset`
+
 # v.0.17.0 (2022-03-23)
 
 ## Update